NPC pathfinding
The Enemy pathfinding has been improved. Now, the enemy will NOT visit the last 5 waypoints its been to, thus making sure that they won't walk around in a constant circle (that was a possibility.) Also removed the need for an enemy health script as i just moved that into enemybehaviour. Added template for ammo pickup (yet to work on but the logic is nearly there). Added player health, and added empty section for logic when the player dies and menu. added two tags, ammo for ammo and visitedwaypoints for the waypoints that have been visited.