Added Level Layout
Added level layout for the prototype. Added 6 enemies to each floor. Added random waypoints to each floor. Did the floors and stairs for Patrick. Added navmesh to all floors and stairs so actors can climb. Sorted everything out in the scene (Added in empty gameobjects to use as lists/seperators)