Updated Prototype
Added player death, added rudimentary light to enemy when they shoot. The player death is instant and *should* just boot to the main menu. The enemy light will shine when they are attacking a player. also modified the primitive enemy to let the player know where the enemy is looking